SherlockAIChatbotWebapp/requirements.txt
absl-py==2.1.0
accelerate==0.33.0
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.2
annotated-types==0.7.0
anyascii==0.3.2
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
Babel==2.15.0
bangla==0.0.2
blinker==1.8.2
blis==0.7.11
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
cachetools==5.4.0
catalogue==2.0.10
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
cmake==3.29.6
colorama==0.4.6
comm==0.2.2
confection==0.1.5
contourpy==1.2.1
coqpit==0.0.17
ctransformers==0.2.27
cuda-python==12.6.0
cycler==0.12.1
cymem==2.0.8
Cython==3.0.10
dateparser==1.1.8
debugpy==1.8.3
decorator==5.1.1
diskcache==5.6.3
docopt==0.6.2
einops==0.8.0
encodec==0.1.1
exceptiongroup==1.2.2
executing==2.0.1
filelock==3.15.4
flash-attn==1.0.4
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.0
g2pkk==0.1.2
google-auth==2.33.0
google-auth-oauthlib==1.2.1
greenlet==3.0.3
grpcio==1.65.1
gruut==2.2.3
gruut-ipa==0.13.0
gruut-lang-de==2.0.1
gruut-lang-en==2.0.1
gruut-lang-es==2.0.1
gruut-lang-fr==2.0.2
hangul-romanize==0.1.0
huggingface-hub==0.24.2
idna==3.7
inflect==7.3.1
intel-openmp==2021.4.0
ipykernel==6.29.5
ipython==8.26.0
itsdangerous==2.2.0
jamo==0.4.1
jedi==0.19.1
jieba==0.42.1
Jinja2==3.1.4
joblib==1.4.2
jsonlines==1.2.0
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
langcodes==3.4.0
language_data==1.2.0
lazy_loader==0.4
librosa==0.10.0
llama_cpp_python==0.2.85
llvmlite==0.43.0
Mako==1.3.5
marisa-trie==1.2.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mkl==2021.4.0
more-itertools==10.3.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==2.8.8
nltk==3.8.1
num2words==0.5.13
numba==0.60.0
numpy==1.22.0
oauthlib==3.2.2
packaging==24.1
pandas==1.5.3
parso==0.8.4
pillow==10.2.0
platformdirs==4.2.2
pooch==1.8.2
preshed==3.0.9
prompt_toolkit==3.0.47
protobuf==4.25.3
psutil==6.0.0
psycopg2-binary==2.9.9
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pynndescent==0.5.13
pyparsing==3.1.2
pypinyin==0.51.0
pyre-extensions==0.0.23
pysbd==0.3.4
python-crfsuite==0.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==26.0.3
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.11.4
seaborn==0.13.2
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
soundfile==0.12.1
soxr==0.3.7
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.31
srsly==2.4.8
stack-data==0.6.3
SudachiDict-core==20240716
SudachiPy==0.6.8
sympy==1.12.1
tbb==2021.13.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
thinc==8.2.5
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.4.0+cu121
torchaudio==2.4.0+cu121
torchvision==0.19.0+cu121
tornado==6.4.1
tqdm==4.66.4
trainer==0.0.36
traitlets==5.14.3
transformers==4.43.2
triton @ file:///C:/Users/786me/Desktop/LLM/triton-2.0.0-cp310-cp310-win_amd64.whl
TTS==0.22.0
typeguard==4.3.0
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
umap-learn==0.5.6
Unidecode==1.3.8
urllib3==2.2.2
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
Werkzeug==3.0.3
wrapt==1.16.0
xformers==0.0.27.post2
yarl==1.9.4
